# openapi.json
{"openapi":"3.1.0","info":{"title":"BlaiseConnect API","version":"0.1.0"},"paths":{"/health":{"get":{"tags":["health"],"summary":"Health Check","description":"Retourne un statut simple indiquant que FastAPI fonctionne.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Check Health Get"}}}}}}},"/auth/login":{"post":{"tags":["authentication"],"summary":"Login","description":"Authentifie un utilisateur et ouvre une session sécurisée.","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["authentication"],"summary":"Get Current Account Information","description":"Retourne les informations du compte actuellement connecté.","operationId":"get_current_account_information_auth_me_get","parameters":[{"name":"blaise_session","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Blaise Session"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CurrentAccountResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/logout":{"post":{"tags":["authentication"],"summary":"Logout","description":"Révoque la session et supprime le cookie du navigateur.","operationId":"logout_auth_logout_post","parameters":[{"name":"blaise_session","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Blaise Session"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/accounts":{"get":{"tags":["accounts"],"summary":"Get Accounts","description":"Retourne les comptes à un administrateur connecté.","operationId":"get_accounts_accounts_get","parameters":[{"name":"blaise_session","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Blaise Session"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountResponse"},"title":"Response Get Accounts Accounts Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["accounts"],"summary":"Post Account","description":"Crée un compte administrateur ou enseignant.","operationId":"post_account_accounts_post","parameters":[{"name":"blaise_session","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Blaise Session"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/school-years":{"get":{"tags":["school-years"],"summary":"Get School Years","description":"Retourne les années scolaires à un administrateur connecté.","operationId":"get_school_years_school_years_get","parameters":[{"name":"blaise_session","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Blaise Session"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SchoolYearResponse"},"title":"Response Get School Years School Years Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["school-years"],"summary":"Post School Year","description":"Crée une nouvelle année scolaire.","operationId":"post_school_year_school_years_post","parameters":[{"name":"blaise_session","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Blaise Session"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchoolYearCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchoolYearResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/school-years/{school_year_id}":{"get":{"tags":["school-years"],"summary":"Get School Year","description":"Retourne le détail d'une année scolaire.","operationId":"get_school_year_school_years__school_year_id__get","parameters":[{"name":"school_year_id","in":"path","required":true,"schema":{"title":"School Year Id"}},{"name":"blaise_session","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Blaise Session"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchoolYearResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/school-years/{school_year_id}/set-current":{"post":{"tags":["school-years"],"summary":"Post Set Current School Year","description":"Définit l'année donnée comme l'unique année courante.","operationId":"post_set_current_school_year_school_years__school_year_id__set_current_post","parameters":[{"name":"school_year_id","in":"path","required":true,"schema":{"title":"School Year Id"}},{"name":"blaise_session","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Blaise Session"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchoolYearResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/school-years/{school_year_id}/close":{"post":{"tags":["school-years"],"summary":"Post Close School Year","description":"Clôture une année scolaire.\n\nFerme automatiquement les inscriptions ouvertes de cette année\n(trigger PostgreSQL) et rend l'année immuable.","operationId":"post_close_school_year_school_years__school_year_id__close_post","parameters":[{"name":"school_year_id","in":"path","required":true,"schema":{"title":"School Year Id"}},{"name":"blaise_session","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Blaise Session"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchoolYearResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/school-years/{school_year_id}/reporting-periods":{"get":{"tags":["reporting-periods"],"summary":"Get Reporting Periods","description":"Retourne les périodes d'une année scolaire, ordonnées par date.","operationId":"get_reporting_periods_school_years__school_year_id__reporting_periods_get","parameters":[{"name":"school_year_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"School Year Id"}},{"name":"blaise_session","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Blaise Session"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReportingPeriodResponse"},"title":"Response Get Reporting Periods School Years  School Year Id  Reporting Periods Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["reporting-periods"],"summary":"Post Reporting Period","description":"Crée une période : le backend calcule sa date de début.","operationId":"post_reporting_period_school_years__school_year_id__reporting_periods_post","parameters":[{"name":"school_year_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"School Year Id"}},{"name":"blaise_session","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Blaise Session"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportingPeriodCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportingPeriodResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/students/":{"get":{"tags":["students"],"summary":"Read Students","description":"Liste et recherche basique des étudiants.","operationId":"read_students_students__get","parameters":[{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Recherche par nom, prénom ou matricule","title":"Q"},"description":"Recherche par nom, prénom ou matricule"},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filtrer par statut : ACTIVE, INACTIVE, ARCHIVED","title":"Status"},"description":"Filtrer par statut : ACTIVE, INACTIVE, ARCHIVED"},{"name":"class_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filtrer par classe id","title":"Class Id"},"description":"Filtrer par classe id"},{"name":"school_year_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filtrer par année scolaire id","title":"School Year Id"},"description":"Filtrer par année scolaire id"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}},{"name":"blaise_session","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Blaise Session"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudentResponse"},"title":"Response Read Students Students  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/students/{student_id}":{"get":{"tags":["students"],"summary":"Read Student","description":"Détail lecture seule d'un élève.","operationId":"read_student_students__student_id__get","parameters":[{"name":"student_id","in":"path","required":true,"schema":{"type":"string","title":"Student Id"}},{"name":"blaise_session","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Blaise Session"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/school-classes":{"get":{"tags":["school-classes"],"summary":"Get School Classes","operationId":"get_school_classes_school_classes_get","parameters":[{"name":"blaise_session","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Blaise Session"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SchoolClassResponse"},"title":"Response Get School Classes School Classes Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AccountCreate":{"properties":{"registration_number":{"type":"string","maxLength":7,"minLength":7,"pattern":"^[aeup][0-9]{6}$","title":"Registration Number"},"password":{"type":"string","maxLength":128,"minLength":8,"format":"password","title":"Password","writeOnly":true},"role":{"type":"string","enum":["ADMIN","TEACHER"],"title":"Role"}},"additionalProperties":false,"type":"object","required":["registration_number","password","role"],"title":"AccountCreate","description":"Données acceptées lors de la création d'un compte."},"AccountResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"registration_number":{"type":"string","maxLength":7,"minLength":7,"pattern":"^[aeup][0-9]{6}$","title":"Registration Number"},"role":{"$ref":"#/components/schemas/AccountRole"},"is_active":{"type":"boolean","title":"Is Active"},"archived_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Archived At"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","registration_number","role","is_active","archived_at","created_at"],"title":"AccountResponse","description":"Informations d'un compte pouvant être retournées par l'API."},"AccountRole":{"type":"string","enum":["ADMIN","TEACHER","STUDENT","GUARDIAN"],"title":"AccountRole","description":"Rôles pouvant être lus et retournés par l'API."},"CurrentAccountResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"registration_number":{"type":"string","maxLength":7,"minLength":7,"pattern":"^[aeup][0-9]{6}$","title":"Registration Number"},"role":{"$ref":"#/components/schemas/AccountRole"},"is_active":{"type":"boolean","title":"Is Active"},"locked_until":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Locked Until"},"last_login_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Login At"},"archived_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Archived At"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","registration_number","role","is_active","locked_until","last_login_at","archived_at","created_at","updated_at"],"title":"CurrentAccountResponse","description":"Informations non sensibles du compte connecté."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"registration_number":{"type":"string","maxLength":7,"minLength":7,"pattern":"^[aeup][0-9]{6}$","title":"Registration Number"},"password":{"type":"string","minLength":1,"format":"password","title":"Password","writeOnly":true}},"additionalProperties":false,"type":"object","required":["registration_number","password"],"title":"LoginRequest","description":"Données envoyées par l'utilisateur pour se connecter."},"LoginResponse":{"properties":{"message":{"type":"string","title":"Message"},"account":{"$ref":"#/components/schemas/CurrentAccountResponse"}},"type":"object","required":["message","account"],"title":"LoginResponse","description":"Réponse retournée après une connexion réussie."},"ReportingPeriodCreate":{"properties":{"school_year_id":{"type":"string","format":"uuid","title":"School Year Id"},"name":{"type":"string","maxLength":100,"minLength":1,"title":"Name"},"end_date":{"type":"string","format":"date","title":"End Date"}},"type":"object","required":["school_year_id","name","end_date"],"title":"ReportingPeriodCreate","description":"Données saisies par l'administrateur pour créer une période."},"ReportingPeriodResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"school_year_id":{"type":"string","format":"uuid","title":"School Year Id"},"name":{"type":"string","title":"Name"},"start_date":{"type":"string","format":"date","title":"Start Date"},"end_date":{"type":"string","format":"date","title":"End Date"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","school_year_id","name","start_date","end_date","created_at"],"title":"ReportingPeriodResponse","description":"Informations d'une période retournées par l'API."},"SchoolClassResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"school_year_id":{"type":"string","format":"uuid","title":"School Year Id"},"name":{"type":"string","title":"Name"},"level":{"type":"string","title":"Level"},"is_active":{"type":"boolean","title":"Is Active"},"archived_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Archived At"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","school_year_id","name","level","is_active","archived_at","created_at","updated_at"],"title":"SchoolClassResponse","description":"Expose les informations d'une classe."},"SchoolYearCreate":{"properties":{"name":{"type":"string","maxLength":20,"minLength":1,"title":"Name"},"start_date":{"type":"string","format":"date","title":"Start Date"},"end_date":{"type":"string","format":"date","title":"End Date"}},"type":"object","required":["name","start_date","end_date"],"title":"SchoolYearCreate","description":"Données nécessaires à la création d'une année scolaire.\n\n`is_current`, `closed_at` et `closed_by_account_id` ne sont pas\nsaisissables à la création : une année démarre toujours ouverte\net non courante par défaut (l'administrateur la définit comme\ncourante via une action dédiée)."},"SchoolYearResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"},"start_date":{"type":"string","format":"date","title":"Start Date"},"end_date":{"type":"string","format":"date","title":"End Date"},"is_current":{"type":"boolean","title":"Is Current"},"closed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Closed At"},"closed_by_account_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Closed By Account Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","name","start_date","end_date","is_current","closed_at","closed_by_account_id","created_at"],"title":"SchoolYearResponse","description":"Informations d'une année scolaire retournées par l'API."},"StudentResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"account_id":{"type":"string","format":"uuid","title":"Account Id"},"registration_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Registration Number"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"birth_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Birth Date"},"gender":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Gender"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"admission_date":{"type":"string","format":"date","title":"Admission Date"},"status":{"type":"string","title":"Status"},"photo_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Photo Path"},"archived_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Archived At"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"class_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Class Id"},"school_year_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"School Year Id"}},"type":"object","required":["id","account_id","registration_number","first_name","last_name","birth_date","gender","email","phone","address","admission_date","status","photo_path","archived_at","created_at","updated_at","class_id","school_year_id"],"title":"StudentResponse","description":"Expose les informations utiles d'un élève."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}